renames for relay lib
