+++ src/libs/relay/c/conduit_relay.h
@@ -44,105 +44,42 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: relay_web_node_viewer_server.hpp
+/// file: conduit_relay.h
 ///
 //-----------------------------------------------------------------------------
 
-#ifndef CONDUIT_RELAY_WEB_NODE_VIEWER_SERVER_HPP
-#define CONDUIT_RELAY_WEB_NODE_VIEWER_SERVER_HPP
+#ifndef CONDUIT_RELAY_H
+#define CONDUIT_RELAY_H
 
 //-----------------------------------------------------------------------------
-// conduit lib includes
-//-----------------------------------------------------------------------------
-#include "conduit.hpp"
-
-#include "relay_exports.hpp"
-#include "relay_web.hpp"
-
-//-----------------------------------------------------------------------------
-// -- begin conduit:: --
-//-----------------------------------------------------------------------------
-namespace conduit
-{
-
-//-----------------------------------------------------------------------------
-// -- begin conduit::relay --
-//-----------------------------------------------------------------------------
-namespace relay 
-{
-
-//-----------------------------------------------------------------------------
-// -- begin conduit::relay::web --
-//-----------------------------------------------------------------------------
-namespace web
-{
-
-//-----------------------------------------------------------------------------
-// -- Viewer Web Request Handler  -
-//-----------------------------------------------------------------------------
-class CONDUIT_RELAY_API NodeViewerRequestHandler : public WebRequestHandler
-{
-public:
-                   NodeViewerRequestHandler();
-                  ~NodeViewerRequestHandler();
-    
-    virtual bool   handle_post(WebServer *server,
-                               struct mg_connection *conn);
-
-    virtual bool   handle_get(WebServer *server,
-                              struct mg_connection *conn);
-                              
-    void           set_node(Node *node);
-
-private:
-    // catch all, used for any post or get
-    bool           handle_request(WebServer *server,
-                                  struct mg_connection *conn);
-    // handlers for specific commands 
-    bool           handle_get_schema(struct mg_connection *conn);
-    bool           handle_get_value(struct mg_connection *conn);
-    bool           handle_get_base64_json(struct mg_connection *conn);
-    bool           handle_shutdown(WebServer *server);
-
-    // holds the node to visualize 
-    Node          *m_node;
-};
-
-//-----------------------------------------------------------------------------
-// -- Node Viewer Web Server -
+// -- includes for the public conduit relay c interface -- 
 //-----------------------------------------------------------------------------
 
-class CONDUIT_RELAY_API NodeViewerServer : public WebServer
-{
-public:
-    
-             NodeViewerServer();
-    virtual ~NodeViewerServer();
+#include "conduit.h"
+#include "conduit_relay_exports.hpp"
 
-    void    set_node(Node *node);
+//-----------------------------------------------------------------------------
+// -- begin extern C
+//-----------------------------------------------------------------------------
+#ifdef __cplusplus
+extern "C" {
+#endif
 
-};
+//-----------------------------------------------------------------------------
+// -- conduit_relay c interface  --
+//-----------------------------------------------------------------------------
 
+CONDUIT_RELAY_API void conduit_relay_about(conduit_node *cnode);
 
+#ifdef __cplusplus
 }
+#endif
 //-----------------------------------------------------------------------------
-// -- end conduit::relay::web --
-//-----------------------------------------------------------------------------
-
-
-}
-//-----------------------------------------------------------------------------
-// -- end conduit::relay --
+// -- end extern C
 //-----------------------------------------------------------------------------
 
 
-
-}
 //-----------------------------------------------------------------------------
-// -- end conduit:: --
+// -- end header guard ifdef
 //-----------------------------------------------------------------------------
-
-#endif 
-
-
-
+#endif
