+++ src/libs/relay/conduit_relay.cpp
@@ -44,22 +44,16 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: relay_silo.hpp
+/// file: conduit_relay.cpp
 ///
 //-----------------------------------------------------------------------------
 
-#ifndef CONDUIT_RELAY_SILO_HPP
-#define CONDUIT_RELAY_SILO_HPP
+#include "conduit_relay.hpp"
 
 //-----------------------------------------------------------------------------
-// external lib includes
+// standard lib includes
 //-----------------------------------------------------------------------------
-#include <silo.h>
-
-//-----------------------------------------------------------------------------
-// conduit includes
-//-----------------------------------------------------------------------------
-#include "relay_io.hpp"
+#include <iostream>
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
@@ -73,57 +67,72 @@ namespace conduit
 namespace relay
 {
 
-//-----------------------------------------------------------------------------
-// -- begin conduit::relay::io --
-//-----------------------------------------------------------------------------
-namespace io
+
+//---------------------------------------------------------------------------//
+std::string
+about()
 {
+    Node n;
+    relay::about(n);
+    return n.to_json();
+}
 
-//-----------------------------------------------------------------------------
-void CONDUIT_RELAY_API silo_write(const Node &node,
-                                  const std::string &path);
+//---------------------------------------------------------------------------//
+void
+about(Node &n)
+{
+    n.reset();
 
-void CONDUIT_RELAY_API silo_read(const std::string &path,
-                                 Node &node);
+    n["web"] = "enabled";
+    
+    Node conduit_about;
+    conduit::about(conduit_about);
+    
+    std::string install_prefix = conduit_about["install_prefix"].as_string();
+    std::string web_root = utils::join_file_path(install_prefix,"share");
+    web_root = utils::join_file_path(web_root,"conduit");
+    web_root = utils::join_file_path(web_root,"web_clients");
+    
+    n["web_client_root"] =  web_root;
 
-//-----------------------------------------------------------------------------
-void CONDUIT_RELAY_API silo_write(const  Node &node,
-                                  const std::string &file_path,
-                                  const std::string &silo_obj_path);
+    Node &io_protos = n["io/protocols"];
 
-void CONDUIT_RELAY_API silo_read(const std::string &file_path,
-                                 const std::string &silo_obj_path,
-                                 Node &node);
+    // standard binary io
+    io_protos["conduit_bin"] = "enabled";
 
-//-----------------------------------------------------------------------------
-void CONDUIT_RELAY_API silo_write(const  Node &node,
-                                  DBfile *dbfile,
-                                  const std::string &silo_obj_path);
+#ifdef CONDUIT_RELAY_IO_HDF5_ENABLED
+    // straight hdf5 
+    io_protos["hdf5"] = "enabled";
+#else
+    // straight hdf5 
+    io_protos["hdf5"] = "disabled";
+#endif
+    
+    // silo
+#ifdef CONDUIT_RELAY_IO_SILO_ENABLED
+    // node is packed into two silo objects
+    io_protos["conduit_silo"] = "enabled";
+#else
+    // node is packed into two silo objects
+    io_protos["conduit_silo"] = "disabled";
+#endif
+    
+    // silo mesh aware
+#ifdef CONDUIT_RELAY_IO_SILO_ENABLED
+    io_protos["conduit_silo_mesh"] = "enabled";
+#else
+    io_protos["conduit_silo_mesh"] = "disabled";
+#endif
 
-void CONDUIT_RELAY_API silo_read(DBfile *dbfile,
-                                 const std::string &silo_obj_path,
-                                 Node &node);
 
-
-//-----------------------------------------------------------------------------    
-void CONDUIT_RELAY_API silo_mesh_write(const Node &mesh,
-                                       const std::string &path);
-
-//-----------------------------------------------------------------------------
-void CONDUIT_RELAY_API silo_mesh_write(const Node &mesh,
-                                       const std::string &file_path,
-                                       const std::string &silo_obj_path);
-
-//-----------------------------------------------------------------------------
-void CONDUIT_RELAY_API silo_mesh_write(const Node &mesh,
-                                       DBfile *dbfile,
-                                       const std::string &silo_obj_path);
+#ifdef CONDUIT_RELAY_MPI_ENABLED
+    n["mpi"] = "enabled";
+#else
+    n["mpi"] = "disabled";
+#endif
 
 
 }
-//-----------------------------------------------------------------------------
-// -- end conduit::relay::io --
-//-----------------------------------------------------------------------------
 
 
 }
@@ -137,6 +146,4 @@ void CONDUIT_RELAY_API silo_mesh_write(const Node &mesh,
 // -- end conduit:: --
 //-----------------------------------------------------------------------------
 
-
-#endif
