+++ src/libs/relay/conduit_relay_config.hpp.in
@@ -44,37 +44,26 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: relay_exports.hpp
+/// file: conduit_relay_config.hpp
 ///
 //-----------------------------------------------------------------------------
 
-#ifndef CONDUIT_RELAY_EXPORTS_HPP
-#define CONDUIT_RELAY_EXPORTS_HPP
+#ifndef CONDUIT_RELAY_CONFIG_HPP
+#define CONDUIT_RELAY_CONFIG_HPP
 
 //-----------------------------------------------------------------------------
-// -- define proper lib exports for various platforms -- 
+//
+// #define optional i/o features
+//
 //-----------------------------------------------------------------------------
-#if defined(_WIN32)
-#if defined(CONDUIT_RELAY_EXPORTS) || defined(conduit_relay_EXPORTS) || defined(CONDUIT_RELAY_MPI_EXPORTS) || defined(conduit_relay_mpi_EXPORTS)
-#define CONDUIT_RELAY_API __declspec(dllexport)
-#else
-#define CONDUIT_RELAY_API __declspec(dllimport)
-#endif
-#if defined(_MSC_VER)
-// Turn off warning about lack of DLL interface
-#pragma warning(disable:4251)
-// Turn off warning non-dll class is base for dll-interface class.
-#pragma warning(disable:4275)
-// Turn off warning about identifier truncation
-#pragma warning(disable:4786)
-#endif
-#else
-# if __GNUC__ >= 4 && ( defined(CONDUIT_RELAY_EXPORTS) || defined(conduit_relay_EXPORTS) || defined(CONDUIT_RELAY_MPI_EXPORTS) || defined(conduit_relay_mpi_EXPORTS))
-#   define CONDUIT_RELAY_API __attribute__ ((visibility("default")))
-# else
-#   define CONDUIT_RELAY_API /* hidden by default */
-# endif
-#endif
+#cmakedefine CONDUIT_RELAY_IO_HDF5_ENABLED
+
+#cmakedefine CONDUIT_RELAY_IO_SILO_ENABLED
+
+#cmakedefine CONDUIT_RELAY_MPI_ENABLED
+
+// this path points to the source dir
+#cmakedefine CONDUIT_RELAY_SOURCE_DIR  "@CONDUIT_RELAY_SOURCE_DIR@"
 
 #endif
