+++ src/libs/relay/conduit_relay_hdf5.cpp
@@ -44,11 +44,11 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: relay_hdf5.cpp
+/// file: conduit_relay_hdf5.cpp
 ///
 //-----------------------------------------------------------------------------
 
-#include "relay_hdf5.hpp"
+#include "conduit_relay_hdf5.hpp"
 
 //-----------------------------------------------------------------------------
 // standard lib includes
