+++ src/libs/relay/conduit_relay_hdf5.hpp
@@ -44,7 +44,7 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: relay_hdf5.hpp
+/// file: conduit_relay_hdf5.hpp
 ///
 //-----------------------------------------------------------------------------
 
@@ -59,7 +59,7 @@
 //-----------------------------------------------------------------------------
 // conduit includes
 //-----------------------------------------------------------------------------
-#include "relay_io.hpp"
+#include "conduit_relay_io.hpp"
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
