+++ src/libs/relay/conduit_relay_io.cpp
@@ -44,11 +44,11 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: relay_io.cpp
+/// file: conduit_relay_io.cpp
 ///
 //-----------------------------------------------------------------------------
 
-#include "relay_io.hpp"
+#include "conduit_relay_io.hpp"
 
 //-----------------------------------------------------------------------------
 // standard lib includes
@@ -57,11 +57,11 @@
 
 // includes for optional features
 #ifdef CONDUIT_RELAY_IO_HDF5_ENABLED
-#include "relay_hdf5.hpp"
+#include "conduit_relay_hdf5.hpp"
 #endif
 
 #ifdef CONDUIT_RELAY_IO_SILO_ENABLED
-#include "relay_silo.hpp"
+#include "conduit_relay_silo.hpp"
 #endif
