+++ src/libs/relay/conduit_relay_mpi.cpp
@@ -44,11 +44,11 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: relay_mpi.cpp
+/// file: conduit_relay_mpi.cpp
 ///
 //-----------------------------------------------------------------------------
 
-#include "relay_mpi.hpp"
+#include "conduit_relay_mpi.hpp"
 #include <iostream>
 
 //-----------------------------------------------------------------------------
