+++ src/libs/relay/conduit_relay_mpi.hpp
@@ -44,7 +44,7 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: relay_mpi.hpp
+/// file: conduit_relay_mpi.hpp
 ///
 //-----------------------------------------------------------------------------
 
@@ -61,7 +61,7 @@
 // conduit includes
 //-----------------------------------------------------------------------------
 #include "conduit.hpp"
-#include "relay_exports.hpp"
+#include "conduit_relay_exports.hpp"
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
