+++ src/libs/relay/conduit_relay_silo.cpp
@@ -44,11 +44,11 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: relay_silo.cpp
+/// file: conduit_relay_silo.cpp
 ///
 //-----------------------------------------------------------------------------
 
-#include "relay_silo.hpp"
+#include "conduit_relay_silo.hpp"
 
 //-----------------------------------------------------------------------------
 // standard lib includes
