+++ src/libs/relay/conduit_relay_web.cpp
@@ -44,7 +44,7 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: relay_web.cpp
+/// file: conduit_relay_web.cpp
 ///
 //-----------------------------------------------------------------------------
 
@@ -62,8 +62,8 @@
 //-----------------------------------------------------------------------------
 // conduit includes
 //-----------------------------------------------------------------------------
-#include "relay_config.hpp"
-#include "relay_web.hpp"
+#include "conduit_relay_config.hpp"
+#include "conduit_relay_web.hpp"
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
