+++ src/libs/relay/python/conduit_relay_io_python.cpp
@@ -64,9 +64,9 @@
 // conduit includes
 //---------------------------------------------------------------------------//
 #include "conduit.hpp"
-#include "relay.hpp"
+#include "conduit_relay.hpp"
 
-#include "Relay_Python_Exports.hpp"
+#include "conduit_relay_python_exports.hpp"
 
 // conduit python module capi header
 #include "conduit_python.hpp"
@@ -357,9 +357,9 @@ static struct PyModuleDef relay_io_python_module_def =
 extern "C" 
 //---------------------------------------------------------------------------//
 #if defined(IS_PY3K)
-PyObject *RELAY_PYTHON_API PyInit_relay_io_python(void)
+PyObject *CONDUIT_RELAY_PYTHON_API PyInit_relay_io_python(void)
 #else
-void RELAY_PYTHON_API initrelay_io_python(void)
+void CONDUIT_RELAY_PYTHON_API initrelay_io_python(void)
 #endif
 //---------------------------------------------------------------------------//
 {
