+++ src/libs/relay/python/conduit_relay_python_exports.hpp
@@ -44,35 +44,40 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: c_relay.cpp
+/// file: conduit_relay_python_exports.hpp
 ///
 //-----------------------------------------------------------------------------
-#include "relay.h"
 
-#include "conduit.hpp"
-#include "relay.hpp"
-#include "conduit_cpp_to_c.hpp"
-
+#ifndef CONDUIT_RELAY_PYTHON_EXPORTS_HPP
+#define CONDUIT_RELAY_PYTHON_EXPORTS_HPP
 
 //-----------------------------------------------------------------------------
-// -- begin extern C
+// -- define proper lib exports for various platforms -- 
 //-----------------------------------------------------------------------------
 
-extern "C" {
+#if defined(_WIN32)
+#if defined(RELAY_PYTHON_EXPORTS) || defined(RELAY_IO_PYTHON_EXPORTS) || defined(relay_python_EXPORTS) || defined(relay_io_python_EXPORTS)
+#define CONDUIT_RELAY_PYTHON_API __declspec(dllexport)
+#else
+#define CONDUIT_RELAY_PYTHON_API __declspec(dllimport)
+#endif
+#if defined(_MSC_VER)
+// Turn off warning about lack of DLL interface
+#pragma warning(disable:4251)
+// Turn off warning non-dll class is base for dll-interface class.
+#pragma warning(disable:4275)
+// Turn off warning about identifier truncation
+#pragma warning(disable:4786)
+#endif
+#else
+# if __GNUC__ >= 4 && (defined(RELAY_PYTHON_EXPORTS) || defined(RELAY_IO_PYTHON_EXPORTS)  || defined(relay_python_EXPORTS) ||  defined(relay_io_python_EXPORTS))
+#   define CONDUIT_RELAY_PYTHON_API __attribute__ ((visibility("default")))
+# else
+#   define CONDUIT_RELAY_PYTHON_API /* hidden by default */
+# endif
+#endif
 
-using namespace conduit;
-
-//---------------------------------------------------------------------------//
-void
-conduit_relay_about(conduit_node *cnode)
-{
-    Node *n = cpp_node(cnode);
-    relay::about(*n);
-}
+#endif
 
 
-}
-//-----------------------------------------------------------------------------
-// -- end extern C
-//-----------------------------------------------------------------------------
