+++ src/libs/relay/python/conduit_relay_web_python.cpp
@@ -60,9 +60,9 @@
 // conduit includes
 //---------------------------------------------------------------------------//
 #include "conduit.hpp"
-#include "relay.hpp"
+#include "conduit_relay.hpp"
 
-#include "Relay_Python_Exports.hpp"
+#include "conduit_relay_python_exports.hpp"
 
 // conduit python module capi header
 #include "conduit_python.hpp"
@@ -963,9 +963,9 @@ static struct PyModuleDef relay_web_python_module_def =
 extern "C" 
 //---------------------------------------------------------------------------//
 #if defined(IS_PY3K)
-PyObject *RELAY_PYTHON_API PyInit_relay_web_python(void)
+PyObject *CONDUIT_RELAY_PYTHON_API PyInit_relay_web_python(void)
 #else
-void RELAY_PYTHON_API initrelay_web_python(void)
+void CONDUIT_RELAY_PYTHON_API initrelay_web_python(void)
 #endif
 //---------------------------------------------------------------------------//
 {
