+++ src/tests/blueprint/t_blueprint_mesh_verify.cpp
@@ -50,7 +50,7 @@
 
 #include "conduit.hpp"
 #include "blueprint.hpp"
-#include "relay.hpp"
+#include "conduit_relay.hpp"
 
 #include <vector>
 #include <string>
