+++ src/tests/relay/t_relay_io_hdf5.cpp
@@ -48,8 +48,8 @@
 ///
 //-----------------------------------------------------------------------------
 
-#include "relay.hpp"
-#include "relay_hdf5.hpp"
+#include "conduit_relay.hpp"
+#include "conduit_relay_hdf5.hpp"
 #include "hdf5.h"
 #include <iostream>
 #include "gtest/gtest.h"
