+++ src/tests/relay/t_relay_io_silo.cpp
@@ -48,8 +48,8 @@
 ///
 //-----------------------------------------------------------------------------
 
-#include "relay.hpp"
-#include "relay_silo.hpp"
+#include "conduit_relay.hpp"
+#include "conduit_relay_silo.hpp"
 #include <iostream>
 #include "gtest/gtest.h"
